+++ packages/plugin-vertical-embed/web/src/typeMapper-loadable.ts
@@ -4,6 +4,6 @@ import loadable from '@loadable/component';
 
 export const typeMapper: PluginTypeMapper = () => ({
   [VERTICAL_EMBED_TYPE]: {
-    component: loadable(() => import('./components/vertical-embed-component')),
+    component: loadable(() => import('./vertical-embed-component')),
   },
 });